Fixed more declaration/include issues and moved Combatant MODULE_ID flags to config structure
