--- src/module/helpers/config.mts
@@ -1,3 +1,0 @@
-export const COSMERE_ADVANCED_ENCOUNTERS: any = {};
-
-// COSMERE_ADVANCED_ENCOUNTERS.attribute = 'advancedencounters.attribute';
